Fix Metrics/ClassLength
Fixes Metrics/ClassLength by extracting `StoryRepository#expand_absolute_urls`
and `StoryRepository#normalize_url` to a separate helper module.
